fixed deadlock when NTP server is installed
When a NTP server is installed, the script cubian-ntpdate will run
forever, and spamming the syslog.
I placed a simple check to the beginning. But probably it should also be
tested if the NTP deamon is running at this moment: e.g. query the
output from /etc/init.d/ntp status